requirements.txt
pandas==2.2.3
torch==2.5.1
pytorch-forecasting==1.2.0
notebook==7.2.2
matplotlib==3.9.2
statsmodels==0.14.4
more-itertools==10.5.0
tensorflow==2.18.0
chronos-forecasting==1.4.1
timesfm==1.2.3
jax==0.4.35
prophet==1.1.6
seaborn==0.13.2
scikit-posthocs==0.11.2
statsannotations==0.7.1
tensorflow-probability==0.25.0
tf-keras==2.18.0
antropy==0.1.9
ruptures==1.1.9
mxnet==1.9.1